migrating some routing files to projects folder
