+++ projects/dendrites/routing/regular_network.py
